+++ src/templates/Post.vue
@@ -1,30 +1,53 @@
 <!-- Post template -->
 <template>
   <Layout class="bg-gray-200 font-body">
-    <!-- <main>
-      <section class="pb-3">
+    <main>
+      <section class="sm:pb-6">
         <div class="max-w-screen-lg mx-auto">
           <header class="mx-6 py-10">
-            <h1 class="font-title font-bold text-4xl">{{ $page.post.title }}</h1>
-            <h2 class="mt-2 text-green-400 text-lg">{{ $page.post.date }}</h2>
-            <h3
-              class="mt-2 text-gray-600 text-lg"
-            >Time to Read: {{ $page.post.timeToRead }} {{ minSuffix }}</h3>
-            <div class="mt-4">
+            <h1 class="font-title font-bold text-3xl md:text-4xl uppercase">{{ $page.post.title }}</h1>
+            <p
+              :class="$page.post.tagCol"
+              class="mt-2 inline-block px-2 py-1 text-sm"
+            >{{ $page.post.tag }}</p>
+            <h2 class="mt-6 text-gray-600 text-lg">
+              {{ $page.post.date }}
+              <span
+                class="ml-6 text-gray-600 text-lg"
+              >{{ $page.post.timeToRead }} {{ minSuffix }}</span>
+            </h2>
+            <div class="mt-6">
               <g-link
                 to="/blog"
-                class="text-lg bg-gray-200 dark:bg-gray-800 rounded p-1 border border-solid border-pink-500 text-pink-500 hover:text-green-400 focus:text-green-400"
-              >Go Back</g-link>
+                class="py-2 px-2 text-lg font-bold border-2 border-solid border-blue-400 hover:border-orange-450 focus:border-orange-450 text-blue-400 hover:text-orange-450 focus:text-orange-450"
+              >More Posts</g-link>
             </div>
           </header>
-          <div class="remark mx-3 bg-gray-100 rounded-md shadow-lg">
-            <article v-html="$page.post.content" class="prose lg:prose-xl mx-auto px-3 py-6"></article>
+          <div class="remark sm:mx-6 bg-gray-100 shadow-lg">
+            <article
+              v-html="$page.post.content"
+              class="prose md:prose-xl mx-auto px-6 sm:px-3 py-10"
+            ></article>
           </div>
         </div>
       </section>
-    </main>-->
+    </main>
   </Layout>
 </template>
+
+<page-query>
+query Post ($path: String!) {
+  post: post (path: $path) {
+    id
+    title
+    date (format: "DD MMMM, YYYY")
+    content
+    timeToRead
+    tag
+    tagCol
+  }
+}
+</page-query>
 
 <script>
 export default {
@@ -36,13 +59,13 @@ export default {
 	methods: {
 		suffixCalc() {
 			if (this.$page.post.timeToRead >= 5) {
-				return (this.minSuffix = 'mins 📖')
+				return (this.minSuffix = 'min read 📖')
 			}
 			if (this.$page.post.timeToRead >= 2) {
-				return (this.minSuffix = 'mins ☕')
+				return (this.minSuffix = 'min read ☕')
 			}
 			if (this.$page.post.timeToRead <= 1) {
-				return (this.minSuffix = 'min 👀')
+				return (this.minSuffix = 'min read 👀')
 			}
 		},
 	},
@@ -54,11 +77,27 @@ export default {
 
 
 <style>
-.remark p code {
-	@apply px-1 rounded;
+.remark h1,
+.remark h2,
+.remark h3,
+.remark h4 {
+	@apply bg-orange-200;
 }
 
-.remark img {
-	@apply rounded-md;
+.remark h1 code,
+.remark h2 code,
+.remark h3 code,
+.remark h4 code,
+.remark p code {
+	@apply px-1;
+}
+
+.remark p a {
+	@apply bg-blue-200 no-underline px-1 transition duration-200 ease-in-out;
+}
+
+.remark p a:hover,
+.remark p a:focus {
+	@apply bg-blue-400;
 }
 </style>